debug

+++ common/op_params.py
@@ -151,13 +151,15 @@ class opParams:
     return param_info.default  # return default value because user's value of key is not in allowed_types to avoid crashing openpilot
 
   def put(self, key, value):
-    lock_file = '/data/op_params.lock'
-    while os.path.exists(lock_file):
-      # print('waiting for lock')
-      time.sleep(1/1000)
-    os.mknod(lock_file)
-    print('no lock!')
-    print(os.path.exists(lock_file))
+    LOCK = False
+    if LOCK:
+      lock_file = '/data/op_params.lock'
+      while os.path.exists(lock_file):
+        # print('waiting for lock')
+        time.sleep(1/1000)
+      os.mknod(lock_file)
+      print('no lock!')
+      print(os.path.exists(lock_file))
 
 
     self._check_key_exists(key, 'put')
@@ -165,7 +167,8 @@ class opParams:
       raise Exception('opParams: Tried to put a value of invalid type!')
     self.params.update({key: value})
     self._write()
-    os.remove(lock_file)
+    if LOCK:
+      os.remove(lock_file)
 
   def delete(self, key):  # todo: might be obsolete. remove?
     if key in self.params:
